+++ src/redturtle/exporter/base/setuphandlers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from App.Common import package_home
+from DateTime import DateTime
 from plone import api
 from plone.app.textfield.value import RichTextValue
 from plone.namedfile.file import NamedBlobImage
@@ -12,7 +13,7 @@ import os
 
 logger = logging.getLogger(__name__)
 
-SIMPLE_TEXT = '''
+SIMPLE_TEXT = """
 <p>Nunc <strong>nulla</strong>. Nullam vel sem. Ut tincidunt tincidunt erat.</p>
 <p>Praesent turpis. Etiam ut purus mattis mauris sodales aliquam.</p>
 <ul>
@@ -20,9 +21,9 @@ SIMPLE_TEXT = '''
 <li>two</li>
 <li>three</li>
 </ul>
-'''
+"""
 
-TEXT_WITH_LINK = '''
+TEXT_WITH_LINK = """
 <p>
   This is an
   <a class="internal-link" href="resolveuid/{uid}" title="">
@@ -33,9 +34,9 @@ TEXT_WITH_LINK = '''
     This is
     <a class="external-link" href="https://www.plone.org" title="">external</a>
 </p>
-'''
+"""
 
-TEXT_WITH_EMPTY_TAGS = '''
+TEXT_WITH_EMPTY_TAGS = """
 <p>Foo</p>
 <p> </p>
 <p><strong><br /></strong></p>
@@ -45,14 +46,14 @@ TEXT_WITH_EMPTY_TAGS = '''
 <p></p>
 <p><i><br /></i></p>
 <i>
-'''
+"""
 
 
 @implementer(INonInstallable)
 class HiddenProfiles(object):
     def getNonInstallableProfiles(self):
         """Hide uninstall profile from site-creation and quickinstaller."""
-        return ['redturtle.exporter.base:default']
+        return ["redturtle.exporter.base:default"]
 
 
 def post_install(context):
@@ -61,78 +62,70 @@ def post_install(context):
     portal = api.portal.get()
 
     # first of all create some contents
-    folder1 = api.content.create(
-        type='Folder', title='Folder foo', container=portal
-    )
-    folder2 = api.content.create(
-        type='Folder', title='Folder bar', container=portal
-    )
-    folder3 = api.content.create(
-        type='Folder', title='Folder baz', container=folder2
-    )
+    folder1 = api.content.create(type="Folder", title="Folder foo", container=portal)
+    folder2 = api.content.create(type="Folder", title="Folder bar", container=portal)
+    folder3 = api.content.create(type="Folder", title="Folder baz", container=folder2)
 
     doc = api.content.create(
-        type='Document',
-        title='First Document',
-        description='Pellentesque habitant morbi tristique senectus',
+        type="Document",
+        title="First Document",
+        description="Pellentesque habitant morbi tristique senectus",
         container=portal,
+        effectiveDate=DateTime(),
     )
     doc2 = api.content.create(
-        type='Document',
-        title='Second Document',
-        description='it\'s inside a folder',
+        type="Document",
+        title="Second Document",
+        description="it's inside a folder",
         container=folder1,
     )
 
     doc3 = api.content.create(
-        type='Document',
-        title='Third document',
-        description='this is the defaulf view of a folder',
+        type="Document",
+        title="Third document",
+        description="this is the defaulf view of a folder",
         container=folder3,
     )
     folder3.setDefaultPage(doc3.getId())
 
     doc4 = api.content.create(
-        type='Document',
-        title='Document with empty tags',
-        description='',
+        type="Document",
+        title="Document with empty tags",
+        description="",
         container=portal,
+        effectiveDate=DateTime(),
     )
 
     news = api.content.create(
-        type='News Item',
-        title='A News',
-        description='In hac habitasse platea dictumst',
+        type="News Item",
+        title="A News",
+        description="In hac habitasse platea dictumst",
         container=portal,
     )
     api.content.create(
-        type='News Item',
-        title='Second News',
-        description='it\'s inside a folder',
+        type="News Item",
+        title="Second News",
+        description="it's inside a folder",
         container=folder2,
     )
 
-    api.content.create(type='Event', title='Event foo', container=portal)
+    api.content.create(type="Event", title="Event foo", container=portal)
 
     api.content.create(
-        type='Collection',
-        title='Collection item',
+        type="Collection",
+        title="Collection item",
         container=portal,
         query=[
             {
-                u'i': u'portal_type',
-                u'o': u'plone.app.querystring.operation.selection.is',
-                u'v': [u'Document', u'News Item'],
+                u"i": u"portal_type",
+                u"o": u"plone.app.querystring.operation.selection.is",
+                u"v": [u"Document", u"News Item"],
             }
         ],
     )
 
-    image = api.content.create(
-        type='Image', title='example image', container=folder3
-    )
-    file_obj = api.content.create(
-        type='File', title='example file', container=folder3
-    )
+    image = api.content.create(type="Image", title="example image", container=folder3)
+    file_obj = api.content.create(type="File", title="example file", container=folder3)
 
     # Now let's add some text and files
     set_text(item=doc, text=SIMPLE_TEXT)
@@ -145,51 +138,47 @@ def post_install(context):
     set_file(item=file_obj)
 
     #  and publish some contents
-    api.content.transition(obj=folder1, transition='publish')
-    api.content.transition(obj=doc, transition='publish')
-    api.content.transition(obj=doc4, transition='publish')
+    api.content.transition(obj=folder1, transition="publish")
+    api.content.transition(obj=doc, transition="publish")
+    api.content.transition(obj=doc4, transition="publish")
 
     #  finally create some users and groups
     api.user.create(
-        username='john',
-        email='jdoe@plone.org',
+        username="john",
+        email="jdoe@plone.org",
         properties=dict(
-            fullname='John Doe',
-            description='foo',
-            home_page='http://www.plone.org',
+            fullname="John Doe", description="foo", home_page="http://www.plone.org"
         ),
     )
-    api.user.create(username='bob', email='bob@plone.org')
-    api.user.grant_roles(username='bob', roles=['Reviewer'])
+    api.user.create(username="bob", email="bob@plone.org")
+    api.user.grant_roles(username="bob", roles=["Reviewer"])
 
-    api.group.create(groupname='staff')
-    group_tool = api.portal.get_tool(name='portal_groups')
-    group_tool.editGroup('staff', roles=['Editor', 'Reader'])
-    api.group.add_user(groupname='Administrators', username='john')
-    api.group.add_user(groupname='staff', username='bob')
+    api.group.create(groupname="staff")
+    group_tool = api.portal.get_tool(name="portal_groups")
+    group_tool.editGroup("staff", roles=["Editor", "Reader"])
+    api.group.add_user(groupname="Administrators", username="john")
+    api.group.add_user(groupname="staff", username="bob")
 
 
-def set_text(item, text, ref=''):
+def set_text(item, text, ref=""):
     if ref:
         text = text.format(uid=ref)
     if IATContentType.providedBy(item):
-        item.setText(text, mimetype='text/html')
+        item.setText(text, mimetype="text/html")
         return
     # dx content
-    item.text = RichTextValue(text, 'text/html', 'text/html')
+    item.text = RichTextValue(text, "text/html", "text/html")
 
 
 def set_image(item):
-    path = os.path.join(package_home(globals()), 'example_files', 'plone.png')
-    with open(path, 'rb') as fd:
+    path = os.path.join(package_home(globals()), "example_files", "plone.png")
+    with open(path, "rb") as fd:
         image_data = fd.read()
-    item.image = NamedBlobImage(data=image_data, filename=u'plone.png')
+    item.image = NamedBlobImage(data=image_data, filename=u"plone.png")
 
 
 def set_file(item):
-    path = os.path.join(
-        package_home(globals()), 'example_files', 'example.pdf'
-    )
-    with open(path, 'rb') as fd:
+    path = os.path.join(package_home(globals()), "example_files", "example.pdf")
+    with open(path, "rb") as fd:
         file_data = fd.read()
-    item.file = NamedBlobImage(data=file_data, filename=u'example.pdf')
+    item.file = NamedBlobImage(data=file_data, filename=u"example.pdf")